LandingPageNav, LargeHeroContent into components
* Move LandingPageNav out of landing-page.js into its own separate file
* Do the same for LargeHeroContent
