+++ pages/landing-page.js
@@ -1,6 +1,8 @@
 // @ts-check
 // eslint-disable-next-line no-unused-vars
 import { ReactElement, useEffect, useState } from 'react'; // ReactElement is only imported for typechecking
+import LandingPageNav from '../components/LandingPageNav';
+import LargeHeroContent from '../components/LargeHeroContent';
 import LargeHeroSection from '../components/LargeHeroSection';
 import Footer from '../components/Footer';
 import Head from 'next/head';
@@ -8,8 +10,6 @@ import Image from 'next/image';
 import Link from 'next/link';
 import PropTypes from 'prop-types';
 
-// import { BsSearch } from 'react-icons/bs';
-import { FaRegHeart } from 'react-icons/fa';
 
 /** -----------------------------------------------------------------------------------------------------------------------
  *
@@ -25,132 +25,11 @@ import { FaRegHeart } from 'react-icons/fa';
 
 
 // -----------------------------------------------------------------------------------------------------------------------
-// TODO: Export LargeHeroContent to components folder.
-// -----------------------------------------------------------------------------------------------------------------------
-
-/**
- * ## Large Hero Content
- * This is the call to action and content of the `LargeHeroSection` component.
- *
- * ### Modify using tailwind classes or css
- * - `titleSize` : "text-`titleSize`" | Example: `"5xl"` | [Tailwind text sizes](https://tailwindcss.com/docs/font-size)
- * - `mainTextSize` : "max-w-`mainSize`" | [Tailwind max-width sizes](https://tailwindcss.com/docs/max-width)
- *
- * @param {Object} props - { mainTextSize, title, titleSize, fadeInSpeed }
- * @param {String} [props.mainTextSize = 'md'] - change the size of main section "max-w-`mainSize`" | [Tailwind max-width sizes](https://tailwindcss.com/docs/max-width)
- * @param {String} [props.fadeInSpeed = '700'] - change the speed of fade in "duration-`fadeInSpeed`" | Example: `700"`
- * @param {String} props.title - The title of the hero section
- * @param {String} [props.titleSize = '5xl'] - "text-`titleSize`" | Example: `"5xl"` | [Tailwind text sizes](https://tailwindcss.com/docs/font-size)
- * @param {*} props.children - The content of the hero section
- * @returns {ReactElement} JSX
- */
-
-const LargeHeroContent = ({
-  children,
-  fadeInSpeed = '700',
-  mainTextSize = 'md',
-  title,
-  titleSize = '5xl',
-}) => {
-  const [opacityAmount, setOpacityAmount] = useState(0);
-  useEffect(() => {
-    setOpacityAmount(1);
-  }, []);
-  return (
-    <div
-      className={`text-center hero-content text-neutral-content opacity-${opacityAmount} transition ease-in duration-${fadeInSpeed}`}
-    >
-      <div className={`max-w-${mainTextSize}`}>
-        <h1 className={`mb-6 text-${titleSize} font-bold`}>{title}</h1>
-        {children}
-      </div>
-    </div>
-  );
-};
-
-LargeHeroContent.propTypes = {
-  children: PropTypes.node,
-  fadeInSpeed: PropTypes.string,
-  mainTextSize: PropTypes.string,
-  title: PropTypes.string.isRequired,
-  titleSize: PropTypes.string,
-};
-
-// -----------------------------------------------------------------------------------------------------------------------
-// Export LargeHeroContent end.
-// -----------------------------------------------------------------------------------------------------------------------
-
-// -----------------------------------------------------------------------------------------------------------------------
 // TODO: Export LandingPageNav to components folder.
 // FIXME: The commented out code is for @Audrey to decide whether or not she wants to use the links in the nav.
 // -----------------------------------------------------------------------------------------------------------------------
 
-/**
- * ### Landing Page Nav Bar
- * This component is used to create the nav bar for the landing page which has opacity and fade in options to work in conjunction with the `LargeHeroSection` component.
- *
- * fadeInSpeed - change the speed of fade in "ease-in duration-`fadeInSpeed`" | Example: `"700"` : [tailwind durations](https://tailwindcss.com/docs/transition-duration)
- *
- * @param {Object} props - { fadeInSpeed }
- * @param {String} [props.fadeInSpeed = '700' ] - change the speed of fade in "ease-in duration-`fadeInSpeed`" | Example: `"700"` : [tailwind durations](https://tailwindcss.com/docs/transition-duration)
- * @returns {ReactElement} JSX
- */
-const LandingPageNav = ({ fadeInSpeed = '700' }) => {
-  const [navOpacity, setnavOpacity] = useState(0);
-  useEffect(() => {
-    setnavOpacity(100);
-  }, []);
-  return (
-    <>
-      <div
-        className={`fixed h-20 nav-bar mb-2 shadow-lg text-neutral-content flex justify-between items-center top-0 left-0 right-0 z-50 opacity-${navOpacity} ease-in duration-${fadeInSpeed}`}
-      >
-        <div className="flex-none px-6 mx-2">
-          <span className="text-lg font-bold">
-            {'Margarita Humanitarian Foundation'}
-          </span>
-        </div>
-        {/* <div className="px-2 mx-2 hidden md:flex flex-grow "> */}
-        {/* <div className="items-stretch hidden lg:flex">
-              <Link href="/">
-                <a className="btn btn-ghost btn-sm rounded-btn">{'Home'}</a>
-              </Link>
-              <Link href="/in-kind">
-                <a className="btn btn-ghost btn-sm rounded-btn">{'In-Kind'}</a>
-              </Link>
-              <Link href="/fund">
-                <a className="btn btn-ghost btn-sm rounded-btn">{'Fund'}</a>
-              </Link>
-              <Link href="/give-your-time">
-                <a className="btn btn-ghost btn-sm rounded-btn">
-                  {'Give Your Time'}
-                </a>
-              </Link>
-              <Link href="/partner-with-us">
-                <a className="btn btn-ghost btn-sm rounded-btn">
-                  {'Partner With Us'}
-                </a>
-              </Link>
-            </div> */}
-        {/* </div> */}
-        <div className="flex items-center">
-          <div className="pr-7 hidden md:block">{' Help Families in Need'}</div>
-          <div className="flex-row hidden sm:flex">
-            <button className="btn btn-square btn-ghost pr-6" type="button">
-              <FaRegHeart className="inline-block w-5 h-5 stroke-current" />
-            </button>
-          </div>
-        </div>
 
-        {/* <div className="flex-none hidden sm:flex">
-            <button className="btn btn-square btn-ghost" type="button">
-              <BsSearch className="inline-block w-5 h-5 stroke-current" />
-            </button>
-          </div> */}
-      </div>
-    </>
-  );
-};
 
 // -----------------------------------------------------------------------------------------------------------------------
 // Export LandingPageNav end.
